Fix CI: collapse match guard and split oversized modules (Fixes #39) (#84)
- Resolve newer-clippy collapsible_if in issues_load_ops.rs via match guard.
- Split files exceeding the 1000-line source policy without behavior change:
  - github/mod.rs -> extract parsing/arg-building into github/parse.rs
  - app_input/mod.rs -> extract issue-list dispatch into issues_list_dispatch.rs
  - state/issues_tests.rs -> split into issues_tests_components.rs
  - state/issues_tests_detail.rs -> split into issues_tests_detail_flow.rs
- Register new test modules in state/mod.rs; keep helpers/imports per file.
