+++ src/state/issues_tests_components.rs
@@ -1,0 +1,337 @@
+use crate::domain::{
+    Issue, IssueComment, IssueDetail, IssueFilter, IssueState, Repository, RepositoryId,
+};
+use crate::state::AppState;
+use crate::state::types::{AppEvent, ComposerTarget, InlineState, IssueFocus};
+
+/// Helper to create a test issue with the given number.
+fn make_test_issue(number: u64) -> Issue {
+    Issue {
+        number,
+        title: format!("Test Issue #{number}"),
+        state: IssueState::Open,
+        author_login: "testuser".to_string(),
+        updated_at: "2024-01-01T00:00:00Z".to_string(),
+        assignee_summary: String::new(),
+        labels_summary: String::new(),
+        comment_count: 0,
+        body: String::new(),
+    }
+}
+
+// -------------------------------------------------------------------------
+// P13 Tests — UI Components + Persistence Rendering Contracts
+// -------------------------------------------------------------------------
+
+/// Helper to build a minimal IssueDetail for testing.
+fn make_test_detail(comments: Vec<IssueComment>) -> IssueDetail {
+    IssueDetail {
+        repo_owner_name: "owner/repo".to_string(),
+        number: 42,
+        title: "Test detail issue".to_string(),
+        state: IssueState::Open,
+        author_login: "octocat".to_string(),
+        created_at: "2024-01-01T00:00:00Z".to_string(),
+        updated_at: "2024-01-02T00:00:00Z".to_string(),
+        labels: vec!["bug".to_string(), "ui".to_string()],
+        assignees: vec!["dev1".to_string()],
+        milestone: Some("v1.0".to_string()),
+        body: "Detail body text".to_string(),
+        external_url: "https://github.com/owner/repo/issues/42".to_string(),
+        comments,
+        has_more_comments: false,
+        comments_cursor: None,
+    }
+}
+
+/// Helper to make a test IssueComment.
+fn make_test_comment(id: u64, author: &str, body: &str) -> IssueComment {
+    IssueComment {
+        comment_id: id,
+        author_login: author.to_string(),
+        created_at: "2024-01-03T00:00:00Z".to_string(),
+        edited_at: None,
+        body: body.to_string(),
+    }
+}
+
+/// Helper to set up a state with a selected repository at index 0.
+fn state_with_repo(repo_id: &str) -> AppState {
+    let mut state = AppState::default();
+    state.repositories.push(Repository::new(
+        RepositoryId(repo_id.to_string()),
+        "Test Repo".to_string(),
+        repo_id.to_string(),
+        std::path::PathBuf::from("/tmp/test-repo"),
+    ));
+    state.selected_repository_index = Some(0);
+    state
+}
+
+/// P13 Test 3: IssueListLoaded with 5 issues populates issues_state.issues with exactly 5 items.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-006
+#[test]
+fn test_issue_list_row_count() {
+    let state = state_with_repo("repo-1").apply(AppEvent::IssueListLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        filter: Box::new(IssueFilter::default()),
+        request_id: 0,
+        issues: (1u64..=5).map(make_test_issue).collect(),
+        cursor: None,
+        has_more: false,
+    });
+
+    assert_eq!(state.issues_state.issues.len(), 5);
+}
+
+/// P13 Test 4: After loading issues and navigating down, selected_issue_index becomes Some(1).
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-006
+#[test]
+fn test_issue_list_selection_highlight() {
+    let state = state_with_repo("repo-1").apply(AppEvent::IssueListLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        filter: Box::new(IssueFilter::default()),
+        request_id: 0,
+        issues: (1u64..=5).map(make_test_issue).collect(),
+        cursor: None,
+        has_more: false,
+    });
+
+    // After load, selection is at 0. Navigate down once.
+    let state = state.apply(AppEvent::IssuesNavigateDown);
+
+    assert_eq!(state.issues_state.selected_issue_index, Some(1));
+}
+
+/// P13 Test 5: Entering issues mode sets list_loading to true initially.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-006
+#[test]
+fn test_issue_list_loading_state() {
+    let state = AppState::default().apply(AppEvent::EnterIssuesMode);
+
+    // list_loading should be true right after EnterIssuesMode (before data arrives)
+    assert!(state.issues_state.loading.list);
+}
+
+/// P13 Test 6: IssueListLoaded with empty vec leaves issues empty and selected_issue_index None.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-006, REQ-ISS-014
+#[test]
+fn test_issue_list_empty_state() {
+    let state = state_with_repo("repo-1").apply(AppEvent::IssueListLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        filter: Box::new(IssueFilter::default()),
+        request_id: 0,
+        issues: vec![],
+        cursor: None,
+        has_more: false,
+    });
+
+    assert!(state.issues_state.issues.is_empty());
+    assert!(state.issues_state.selected_issue_index.is_none());
+}
+
+/// P13 Test 7: IssueDetailLoaded populates all fields in issues_state.issue_detail.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-009
+#[test]
+fn test_issue_detail_all_fields() {
+    let comments = vec![make_test_comment(1, "alice", "Looks good")];
+    let detail = make_test_detail(comments);
+
+    let mut state = state_with_repo("repo-1");
+    state.mark_issue_detail_loading(RepositoryId("repo-1".to_string()), 42);
+    let state = state.apply(AppEvent::IssueDetailLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        issue_number: 42,
+        request_id: 0,
+        detail: Box::new(detail),
+    });
+
+    let loaded = state
+        .issues_state
+        .issue_detail
+        .unwrap_or_else(|| panic!("detail should be Some"));
+    assert_eq!(loaded.number, 42);
+    assert_eq!(loaded.title, "Test detail issue");
+    assert_eq!(loaded.author_login, "octocat");
+    assert_eq!(loaded.body, "Detail body text");
+    assert_eq!(loaded.labels, vec!["bug".to_string(), "ui".to_string()]);
+    assert_eq!(loaded.assignees, vec!["dev1".to_string()]);
+    assert_eq!(loaded.milestone, Some("v1.0".to_string()));
+    assert!(!loaded.external_url.is_empty());
+    assert_eq!(loaded.repo_owner_name, "owner/repo");
+}
+
+/// P13 Test 8: IssueDetailLoaded with 3 comments — detail.comments.len() == 3.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-009
+#[test]
+fn test_issue_detail_comments_timeline() {
+    let comments = vec![
+        make_test_comment(10, "alice", "First"),
+        make_test_comment(11, "bob", "Second"),
+        make_test_comment(12, "carol", "Third"),
+    ];
+    let detail = make_test_detail(comments);
+
+    let mut state = state_with_repo("repo-1");
+    state.mark_issue_detail_loading(RepositoryId("repo-1".to_string()), 42);
+    let state = state.apply(AppEvent::IssueDetailLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        issue_number: 42,
+        request_id: 0,
+        detail: Box::new(detail),
+    });
+
+    let loaded = state
+        .issues_state
+        .issue_detail
+        .unwrap_or_else(|| panic!("detail should be Some"));
+    assert_eq!(loaded.comments.len(), 3);
+    assert_eq!(loaded.comments[0].author_login, "alice");
+    assert_eq!(loaded.comments[2].author_login, "carol");
+}
+
+/// P13 Test 9: OpenNewCommentComposer transitions inline_state to Composer(NewComment).
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-010
+#[test]
+fn test_issue_detail_inline_composer_visible() {
+    let mut state = AppState::default();
+    state.issues_state.inline_state = InlineState::None;
+
+    let state = state.apply(AppEvent::OpenNewCommentComposer);
+
+    match state.issues_state.inline_state {
+        InlineState::Composer {
+            target: ComposerTarget::NewComment,
+            ..
+        } => {} // Correct
+        other => panic!("expected Composer(NewComment), got {other:?}"),
+    }
+}
+
+/// P13 Test 9b: OpenNewIssueComposer transitions inline_state to Composer(NewIssue).
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-010
+#[test]
+fn test_issue_list_new_issue_composer_visible() {
+    let mut state = AppState::default();
+    state.issues_state.inline_state = InlineState::None;
+    state.issues_state.issue_focus = IssueFocus::IssueDetail;
+
+    let state = state.apply(AppEvent::OpenNewIssueComposer);
+
+    match state.issues_state.inline_state {
+        InlineState::Composer {
+            target: ComposerTarget::NewIssue,
+            ..
+        } => {}
+        other => panic!("expected Composer(NewIssue), got {other:?}"),
+    }
+    assert_eq!(state.issues_state.issue_focus, IssueFocus::IssueList);
+}
+
+/// P13 Test 10: UpdateDraftFilter sets values in draft_filter fields.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-008
+#[test]
+fn test_filter_controls_value_binding() {
+    let mut state = AppState::default();
+    state.issues_state.filter_ui.controls_open = true;
+
+    // Update multiple draft filter fields
+    let state = state
+        .apply(AppEvent::UpdateDraftFilter {
+            field: "author".to_string(),
+            value: "octocat".to_string(),
+        })
+        .apply(AppEvent::UpdateDraftFilter {
+            field: "assignee".to_string(),
+            value: "dev1".to_string(),
+        })
+        .apply(AppEvent::UpdateDraftFilter {
+            field: "query_text".to_string(),
+            value: "segfault".to_string(),
+        });
+
+    assert_eq!(state.issues_state.draft_filter.author, "octocat");
+    assert_eq!(state.issues_state.draft_filter.assignee, "dev1");
+    assert_eq!(state.issues_state.draft_filter.query_text, "segfault");
+}
+
+/// P13 Test 11: Loading an empty issue list means the empty-state condition holds
+/// (issues.is_empty() is the data contract the UI component checks).
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-014
+#[test]
+fn test_empty_state_no_issues() {
+    let state = state_with_repo("repo-1").apply(AppEvent::IssueListLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        filter: Box::new(IssueFilter::default()),
+        request_id: 0,
+        issues: vec![],
+        cursor: None,
+        has_more: false,
+    });
+
+    // The UI rendering component checks this condition to show the empty message
+    assert!(state.issues_state.issues.is_empty());
+    assert!(!state.issues_state.loading.list);
+}
+
+/// P13 Test 12: IssueDetailLoaded with no comments — detail.comments is empty.
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-014
+#[test]
+fn test_empty_state_no_comments() {
+    let detail = make_test_detail(vec![]);
+
+    let mut state = state_with_repo("repo-1");
+    state.mark_issue_detail_loading(RepositoryId("repo-1".to_string()), 42);
+    let state = state.apply(AppEvent::IssueDetailLoaded {
+        scope_repo_id: RepositoryId("repo-1".to_string()),
+        issue_number: 42,
+        request_id: 0,
+        detail: Box::new(detail),
+    });
+
+    let loaded = state
+        .issues_state
+        .issue_detail
+        .unwrap_or_else(|| panic!("detail should be Some"));
+    assert!(loaded.comments.is_empty());
+}
+
+/// P13 Test 13: OpenAgentChooser with no agents leaves agent_chooser as None
+/// (UI empty-state: no agents available to send to).
+///
+/// @plan PLAN-20260329-ISSUES-MODE.P13
+/// @requirement REQ-ISS-014
+#[test]
+fn test_empty_state_no_agents_for_send() {
+    let mut state = AppState::default();
+    // Confirm no agents are configured
+    assert!(state.agents.is_empty());
+
+    // OpenAgentChooser with no agents should leave chooser as None
+    state = state.apply(AppEvent::OpenAgentChooser);
+
+    // When agents list is empty, agent_chooser is not opened
+    assert!(state.issues_state.agent_chooser.is_none());
+}
